src/surquest/GCP/dataform-cli/requirements.txt
google-auth=2.40.3
google-auth-oauthlib=1.2.2
google-api-python-client=2.173.0